refactor: Makes useInfiniteScrollResults hook - #41
